+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added integration of the new AMP API field `mode`.
+- Added possibility to render embed banners.
+- Added new rendering mode `embed` (`EmbedRenderingMode`) for the Latte bridge.
+
+### Changed
+- Updated docs
 
 ## 1.1.0 - 2023-12-01
 ### Added